% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{choose_readlengths}
\alias{choose_readlengths}
\title{Filter read lengths for P-sites position calculation}
\usage{
choose_readlengths(summary_data, choice = "max_coverage", nt_signals)
}
\arguments{
\item{summary_data}{output data from the \code{calc_cutoffs_from_profiles} function}

\item{choice}{Method used to select readlengths, defaults to "max_coverage"}

\item{nt_signals}{Profiles of 5'ends around start codons}
}
\value{
a list object containing different compartments. Each sub-list contains \code{final_choice}, 
the set of chosen read lengths with cutoffs, and \code{data}, the complete stats for each selection method
}
\description{
This function selects a subset of readlenghts to be used in the P-sites calculation step
}
\details{
Three different methods are available to choose readlengths: the "max_coverage" method
selects all read lenghts with more in-frame signal compared to out-of-frame signal, on all codons;
the "max_inframe" method starts with the most accurate read length and
progressively selects read lengths which add in-frame signals in codons not covered by previous read lengths;
the "all" method selects all available read lengths
}
\seealso{
\code{\link{calc_cutoffs_from_profiles}}
}
\author{
Lorenzo Calviello, \email{calviello.l.bio@gmail.com}
}
\keyword{Ribo-seQC}
