% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{create_html_report}
\alias{create_html_report}
\title{Create the RiboseQC analysis report in html}
\usage{
create_html_report(
  input_files,
  input_sample_names,
  output_file,
  extended = FALSE
)
}
\arguments{
\item{input_files}{Character vector with full paths to data files
generated with \code{RiboseQC_analysis}.
Must be of same length as \code{input_sample_names}.}

\item{input_sample_names}{Character vector containing input names (max. 5 characters per name).
Must be of same length as \code{input_files}.}

\item{output_file}{String; full path to html report file.}

\item{extended}{creates a large html report including codon occupancy for each read length. Defaults to \code{FALSE}}
}
\value{
The function saves the html report file with the file path \code{output_file} and
a folder containing all figures shown in the html report as RDS object files (located in the same directory as the html report).
}
\description{
This function creates the RiboseQC html report based on the RiboseQC analysis files
generated with \code{RiboseQC_analysis}.
}
\details{
This function creates the html report visualizing the RiboseQC analysis data. \cr \cr
Input are two lists of the same length: \cr \cr
a) \code{input_files}: list of full paths to one or multiple input files
(RiboseQC analysis files generated with \code{RiboseQC_analysis}) and \cr \cr
b) \code{input_sample_names}: list of corresponding names describing the file content in max. 5 characters
(these are used as names in the report). \cr \cr
For the report, a RMarkdown file is rendered as html document, saved as \code{output_file}. \cr \cr
Additionally, all figures in the report are saved as PDF figures
in an extra folder in the same directory as the report html file. \cr \cr
Example: \cr
\code{output_file <- "\\mydir\\myreport.html"} will generate
the html report \code{\\mydir\\myreport.html} and
the folder \code{\\mydir\\myreport_plots\\} for the RDS object files to be stored in.
}
\examples{
\dontrun{
sampnames = c("root","shoot")
create_html_report(input_files=paste(sampnames,"_results_RiboseQC",sep = ""), input_sample_names=sampnames,output_file =  'test_root_shoots.html',extended = FALSE)
}
}
\seealso{
\code{\link{plot_read_biotype_dist_1}}, \code{\link{plot_read_biotype_dist_2}},
\code{\link{plot_read_length_dist}}, \code{\link{plot_read_length_dist_by_biotype}},
\code{\link{plot_read_biotype_dist_by_length}},
\code{\link{get_metagene_data}},
\code{\link{plot_metagene_hm_rmd}}, \code{\link{plot_metagene_hm}},
\code{\link{plot_metagene_bar_rmd}}, \code{\link{plot_metagene_bar}},
\code{\link{plot_frame_dist_boxplot_rmd}}, \code{\link{plot_frame_dist_boxplot}},
\code{\link{get_rl_and_cutoffs}}, \code{\link{get_default_rl_selection}},
\code{\link{get_top50_mapping}}, \code{\link{get_top50_cds_genes}}, \code{\link{get_top50_all_genes}},
\code{\link{get_codon_usage_data}},
\code{\link{plot_codon_usage_positional_rmd}}, \code{\link{plot_codon_usage_positional}},
\code{\link{plot_codon_usage_bulk_rmd}}, \code{\link{plot_codon_usage_bulk}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{RiboseQC}
