% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{create_pdfs_from_rds_objects}
\alias{create_pdfs_from_rds_objects}
\title{Generate PDF files from RDS object files}
\usage{
create_pdfs_from_rds_objects(output_rds_path)
}
\arguments{
\item{output_rds_path}{String; full path to output folder for RDS object files. Example: /my_path_to/rds/}
}
\value{
This function creates PDF files from RDS object files.
}
\description{
This function generates figures as PDF files from RDS object files -
it's essentially a workaround for  difficulty integrating pdfs into
html reports.
}
\examples{
create_pdfs_from_rds_objects(paste('test_root_shoots.html',"_plots/rds/",sep=""))
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{RiboseQC}
