% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{get_codon_usage_data}
\alias{get_codon_usage_data}
\title{Get codon usage data (positional and bulk)}
\usage{
get_codon_usage_data(data, data_type, comp, rl)
}
\arguments{
\item{data}{Object (list of lists) generated by \code{RiboseQC_analysis}: \code{res_all}.}

\item{data_type}{String; select one of the following: \cr
\itemize{
\item \code{Codon_counts}: codon count in defined positions*, \cr
\item \code{P_sites_percodon}: P-sites count in defined positions*, or \cr
\item \code{P_sites_percodon_ratio}: ratio of P-sites counts to codon counts in defined positions*.
\item \code{E_sites_percodon}: E-sites count in defined positions*, or \cr
\item \code{E_sites_percodon_ratio}: ratio of E-sites counts to codon counts in defined positions*.
\item \code{A_sites_percodon}: A-sites count in defined positions*, or \cr
\item \code{A_sites_percodon_ratio}: ratio of A-sites counts to codon counts in defined positions*.
}}

\item{comp}{String for originating compartment. \cr
Check for available originating compartments in the data set using: \cr
\code{names(res_all$profiles_P_sites$Codon_count)}}

\item{rl}{String for read length. \cr
Check for available read lengths in the data set using: \cr
\code{names(res_all$profiles_P_sites[[data_type]][[comp]])}}
}
\value{
This function returns a list (e.g. called codon_usage_bulk_data) with information on bulk codon usage: \cr
\itemize{
\item \code{codon_usage_bulk_data$data} contains the data on bulk codon usage,
\item \code{codon_usage_bulk_data$data_type} saves the data type used (see parameter \code{data_type})
\item \code{codon_usage_bulk_data$comp} saves the originating compartment used (see parameter \code{comp}), and
\item \code{codon_usage_bulk_data$rl} saves the read length used (see parameter \code{rl}).
}
}
\description{
This function processes codon usage data generated by \code{RiboseQC_analysis}, i.e.
codon usage 
\itemize{
\item per nucleotide position (i.e. positional codon usage) as well as
\item summed up over all positions* (i.e. bulk codon usage)
}
for a specific data type, originating compartment, and read length, as well as based on a user-defined genetic code. \cr \cr
This data is used as input in \code{plot_codon_usage_bulk} 
to generate a bar plot, and in \code{plot_codon_usage_bulk_rmd} to iterativly generate plots for the Ribo-seQC report in section 7.2. \cr\cr
Please check \code{plot_codon_usage} for positional codon usage (instead of bulk codon usage). \cr\cr
* Information on positions included in analysis: \cr
Based on P-site corrected reads, the codon usage within CDS regions for protein coding genes 
is examplatory calculated 
\itemize{
\item for the first 11 codons of the CDS (referred to as \emph{start}), 
\item for 11 codons from the middle of the CDS (referred to as \emph{middle}), and 
\item for the last 11 codons of the CDS (referred to as \emph{stop}). 
}
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
