% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{get_default_rl_selection}
\alias{get_default_rl_selection}
\title{Get default choice of read lengths}
\usage{
get_default_rl_selection(rdata_list)
}
\arguments{
\item{rdata_list}{List of RiboseQC analysis RData objects generated by \code{generate_rdata_list}.}
}
\value{
This function returns data.
}
\description{
This function returns selected read lengths per originating compartment. 
These read lengths build the basis for all P-site based calculations such as
metagene and codon usage analyses 
(e.g. \code{plot_metagene_hm} and \code{plot_codon_usage}) \cr \cr
This data is used in the Ribo-seQC report in section 4.2.3 (displayed as table).
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
