% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{get_metagene_data}
\alias{get_metagene_data}
\title{Get 5'/P-site profile data for metagene analysis}
\usage{
get_metagene_data(data, profile_type, res, comp)
}
\arguments{
\item{res}{Resolution (subcodon or bins)  \cr \cr
\bold{Subcodon} resolution: \cr
\code{five_prime_subcodon} \cr 
(in order to call \code{res_all$profiles_fivepr$five_prime_subcodon}) or \cr 
\code{P_sites_subcodon} \cr 
(in order to call \code{res_all$profiles_P_sites$P_sites_subcodon}) \cr \cr
Read coverage for 
the first 25nt after the transcription start site (TSS), 
25nt before and 33nt after the start codon, 
33nt from the middle of the CDS, 
33nt before and 25nt after the stop codon, 
and the last 25nt before the transcription end site (TES). \cr \cr
\bold{Bins}: \cr
\code{five_prime_bins} \cr
(in order to call \code{res_all$profiles_fivepr$five_prime_bins}) or \cr
\code{P_sites_bins} \cr
(in order to call \code{res_all$profiles_P_sites$P_sites_bins}) \cr \cr
Read coverage for 
50 bins between TSS and start codon, 100 bins for the CDS, and 50 after stop codon to TES.}

\item{comp}{String for originating compartment \cr \cr
Check for available originating compartments in the data set using:
\code{names(res_all$profiles_fivepr$five_prime_subcodon)}}

\item{profiles}{5' or P-site profile data generated by \code{RiboseQC_analysis}: \cr \cr
\code{res_all$profiles_fivepr} or \cr
\code{res_all$profiles_P_sites} \cr \cr
Consists of DataFrames each 
containing counts of 5' or P-site profiles, 
calculated for different resolution types (see parameter \code{res}), 
originating compartments (see parameter \code{comp}), and
read lengths per input sample. \cr \cr
Example to access DataFrame: \cr 
\code{res_all$profiles_fivepr[[res]][[comp]][[read_length]]} or \cr
\code{res_all$profiles_P_sites[[res]][[comp]][[read_length]]}}
}
\value{
This function returns data \code{profile_data} 
as \code{list(data_single, data_all, res)} with profile data
\itemize{
\item for all read lengths individually (\code{profile_data[1]]} is \code{data_single}) and
\item for all read lenghts summarized (\code{profile_data[[2]]} is \code{data_all}).
}
with different scaling: no scaling (\code{none}), log2 scaling (\code{log2}), and z scoring (\code{zscore}), 
accessable via e.g. \code{profile_data[[1]]$none} or \code{profile_data[[2]]$zscore}. \cr \cr

\code{profile_data[[3]]} saves the resolution type \code{res} for later use during plotting. \cr \cr

\code{profile_data[[4]]} stores information on whether data represents 5' or P site profiles (\code{profile_type}) 
for later use during plotting.
}
\description{
This function processes profile data generated by \code{RiboseQC_analysis}. \cr \cr
This data is used as input in \code{plot_metagene_hm} to generate plot for the Ribo-seQC report in section 4.1/4.3.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
