% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{get_rl_and_cutoffs}
\alias{get_rl_and_cutoffs}
\title{Get selected read lengths and cutoffs}
\usage{
get_rl_and_cutoffs(rdata_list)
}
\arguments{
\item{rdata_list}{List of RiboseQC analysis RData objects generated by \code{generate_rdata_list}.}
}
\value{
This function returns data.
}
\description{
This function retrieves data on selected read lengths (per originating compartment),
e.g. their cutoff values, frame preference and codon gain. \cr \cr
This data is used in the Ribo-seQC report in section 4.2.2 (displayed as table).
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
