% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{get_top50_all_genes}
\alias{get_top50_all_genes}
\title{Get top 50 abundant genes (all genes)}
\usage{
get_top50_all_genes(rdata_list)
}
\arguments{
\item{rdata_list}{List of RiboseQC analysis RData objects generated by \code{generate_rdata_list}.}
}
\value{
This function returns data to be displayed as table in the html report.
}
\description{
This function retrieves data on the top 50 abundant genes. \cr \cr
This data is used in the Ribo-seQC report in section 6 (displayed as table).
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
