% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_codon_usage_bulk}
\alias{plot_codon_usage_bulk}
\title{Plot bulk codon usage bar plots}
\usage{
plot_codon_usage_bulk(codon_usage_data, sample = "", output_rds_path = "")
}
\arguments{
\item{sample}{String; sample name (selected from the input names
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}

\item{codon_usage_bulk_data}{List containing codon usage bulk data and meta data, generated by \code{get_codon_usage_data}.}
}
\value{
This function returns a plot that can be integrated in the html report and
that can be saved as RDS object file.
}
\description{
This function plots the codon usage summed up over all positions* (bulk codon usage) as bar plot
for a specific data type, originating compartment, and read length, as well as based on a user-defined genetic code. \cr \cr
* Information on positions included in analysis: \cr
Based on P-site corrected reads, the codon usage within CDS regions for protein coding genes
is examplatory calculated
\itemize{
\item for the first 11 codons of the CDS (referred to as \emph{start}),
\item for 11 codons from the middle of the CDS (referred to as \emph{middle}), and
\item for the last 11 codons of the CDS (referred to as \emph{stop}).
}
}
\examples{
data(res_all)
plot_codon_usage_bulk_rmd(res_all, 'shoot', paste0('myfigpath', "rds/"))
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{RiboseQC}
