% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_codon_usage_bulk_rmd}
\alias{plot_codon_usage_bulk_rmd}
\title{Plot bulk codon usage bar plots within the RMarkdown document}
\usage{
plot_codon_usage_bulk_rmd(data, sample = "", output_rds_path = "")
}
\arguments{
\item{data}{Object (list of lists) generated by \code{RiboseQC_analysis}: \code{res_all}.}

\item{sample}{String; sample name (selected from the input names 
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
This function returns iteratively all bulk codon usage plots 
for the html report and saves the same plots as RDS object file.
}
\description{
This function generates iteratively all bulk codon usage bar plots; 
iteration over originating compartments, read length, and data type (codon count, read count, codon-read count ratio). \cr \cr
These plots are displayed in the Ribo-seQC report in section 8.
}
\seealso{
\code{\link{create_html_report}}, \code{\link{get_metagene_data}}, \code{\link{plot_metagene_hm}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
