% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_codon_usage_positional_rmd}
\alias{plot_codon_usage_positional_rmd}
\title{Plot positional codon usage heatmaps within the RMarkdown document}
\usage{
plot_codon_usage_positional_rmd(data, sample = "", output_rds_path = "")
}
\arguments{
\item{data}{Object (list of lists) generated by \code{RiboseQC_analysis}: \code{res_all}.}

\item{sample}{String; sample name (selected from the input names
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
This function returns iteratively all positional codon usage plots
for the html report and saves the same plots as RDS object file.
}
\description{
This function generates iteratively all positional codon usage heatmaps;
iteration over originating compartments, read length, data type (codon count, read count, codon-read count ratio),
and scaling method (none, log2, zscale). \cr \cr
These plots are displayed in the RiboseQC report in section 7.
}
\examples{
data(res_all)
plot_codon_usage_positional_rmd(res_all, 'shoot', paste0('myfigpath', "rds/"))
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{RiboseQC}
