% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_metagene_bar}
\alias{plot_metagene_bar}
\title{Plot 5'/P-site profile per read length as barplot}
\usage{
plot_metagene_bar(metagene_data, rl, sample = "", output_rds_path = "")
}
\arguments{
\item{sample}{String; sample name (selected from the input names
iven in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}

\item{data}{Profile data for a resolution type, specific originating compartment, and read length. \cr \cr
Example: \cr
\code{\strong{res_all$profiles_fivepr}$five_prime_subcodon$nucl$'30'} \cr or \cr
\code{\strong{res_all$profiles_P_sites}$five_prime_subcodon$nucl$'30'} \cr \cr
Shown in bold are fixed list names, remaining list names should be adapted to 
the resolution type, specific originating compartment, and read length of interest.}
}
\value{
This function returns a plot that can be integrated in the html report and that can be saved as RDS object file.
}
\description{
This function plots a 5' or P-site profile as barplot
(for a specific originating compartment, resolution type, and read length). \cr \cr
This plot is used in the Ribo-seQC report in section 4.1/4.3.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
