% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_metagene_bar_rmd}
\alias{plot_metagene_bar_rmd}
\title{Plot 5'/P-site profile per read length as barplot in R Markdown}
\usage{
plot_metagene_bar_rmd(metagene_data, sample = "", output_rds_path = "")
}
\arguments{
\item{metagene_data}{...}

\item{sample}{String; sample name (selected from the input names given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function. Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
...
}
\description{
This function plots a 5' or P-site profile as barplot
(for a specific originating compartment, resolution type, and read length). \cr \cr
This plot is used in the Ribo-seQC report in section 4.1/4.3.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
