% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_metagene_hm}
\alias{plot_metagene_hm}
\title{Plot 5'/P-site profiles (for all read lengths) as heatmap}
\usage{
plot_metagene_hm(metagene_data, scal, sample = "", output_rds_path = "")
}
\arguments{
\item{scal}{Scaling method:
no scaling (\code{none}), log2 scaling (\code{log2}), or z scoring (\code{zscore}).}

\item{sample}{String; sample name (selected from the input names
iven in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}

\item{data_profile}{Profile data for a specific originating compartment and resolution type,
generated using \code{\link{get_metagene_data}}.}
}
\value{
This function returns a plot that can be integrated in the html report and  #' that can be saved as RDS object file.
}
\description{
This function plots a 5' or P-site profiles for all read lengths as heatmap
(for a specific originating compartment, resolution type, and scaling method). \cr \cr
This plot is used in the RiboseQC report in section 4.1/4.3.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{RiboseQC}
