% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_metagene_hm_rmd}
\alias{plot_metagene_hm_rmd}
\title{Plot 5'/P-site profiles as heatmaps within the RMarkdown document}
\usage{
plot_metagene_hm_rmd(data, profile_type, sample = "", output_rds_path = "")
}
\arguments{
\item{sample}{String; sample name (selected from the input names
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}

\item{profiles}{5' or P-site profile data generated by \code{RiboseQC_analysis}: \cr \cr
\code{res_all$profiles_fivepr} or \cr
\code{res_all$profiles_P_sites} \cr \cr
Consists of DataFrames each
containing counts of 5' or P-site profiles,
calculated for different resolution types (see parameter \code{res}),
originating compartments (see parameter \code{comp}), and
read lengths per input sample. \cr \cr
Example to access DataFrame: \cr
\code{res_all$profiles_fivepr[[res]][[comp]][[read_length]]} or \cr
\code{res_all$profiles_P_sites[[res]][[comp]][[read_length]]}}
}
\value{
This function returns iteratively all 5' or P-site profile plots
for the html report and saves the same plots as RDS object file.
}
\description{
This function generates iteratively all heatmap plots
(iteration over originating compartments, resolution types, and scaling methods). \cr \cr
These plots are displayed in the RiboseQC report in section 4.1/4.3.
}
\examples{
data(res_all)
plot_metagene_hm_rmd(res_all, "profiles_fivepr", 'shoot', paste0('myfigpath', "rds/"))
}
\seealso{
\code{\link{create_html_report}}, \code{\link{get_metagene_data}}, \code{\link{plot_metagene_hm}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{RiboseQC}
