% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_read_biotype_dist_1}
\alias{plot_read_biotype_dist_1}
\title{Plot read location distribution by biotype (and originating compartment)}
\usage{
plot_read_biotype_dist_1(pos, sample, output_rds_path = "")
}
\arguments{
\item{pos}{\code{res_all$read_stats$positions} generated by \code{RiboseQC_analysis}. \cr \cr
\code{data.frame} containing the number of reads per biotype (rows) and originating compartment (columns).}

\item{sample}{String; sample name (selected from the input names 
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
This function returns a plot that can be integrated in the html report and
that can be saved as RDS object file.
}
\description{
This function plots the read location distribution by biotype (and originating compartment)
for one input sample. \cr \cr
This plot is used in the Ribo-seQC report in section 1.1.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
