% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_read_biotype_dist_2}
\alias{plot_read_biotype_dist_2}
\title{Plot read location distribution by originating compartment (and biotype)}
\usage{
plot_read_biotype_dist_2(rdata_list, output_rds_path = "")
}
\arguments{
\item{rdata_list}{List of RiboseQC analysis RData objects generated by \code{generate_rdata_list}.}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
This function returns a plot that can be integrated in the html report and
that can be saved as RDS object file.
}
\description{
This function plots the read location distribution by originating compartment (and biotype)
for all input samples. \cr \cr 
This plot is used in the Ribo-seQC report in section 1.2.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
