% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{prepare_annotation_files}
\alias{prepare_annotation_files}
\title{Prepare comprehensive sets of annotated genomic features}
\usage{
prepare_annotation_files(
  annotation_directory,
  twobit_file = NULL,
  gtf_file,
  scientific_name = "Homo.sapiens",
  annotation_name = "genc25",
  export_bed_tables_TxDb = TRUE,
  forge_BSgenome = FALSE,
  genome_seq = NULL,
  circ_chroms = DEFAULT_CIRC_SEQS,
  create_TxDb = TRUE,
  annot_file = NULL
)
}
\arguments{
\item{annotation_directory}{The target directory which will contain the output files}

\item{twobit_file}{Full path to the genome file in twobit format}

\item{gtf_file}{Full path to the annotation file in GTF format}

\item{scientific_name}{A name to give to the organism studied; must be two words separated by a ".", defaults to Homo.sapiens}

\item{annotation_name}{A name to give to annotation used; defaults to genc25}

\item{export_bed_tables_TxDb}{Export coordinates and info about different genomic regions in the annotation_directory? It defaults to \code{TRUE}}

\item{forge_BSgenome}{Forge and install a \code{BSgenome} package? It defaults to \code{TRUE}}

\item{genome_seq}{Full path to the genome in FASTA format, used when \code{forge_BSgenome} is \code{FALSE}}

\item{circ_chroms}{Names of the sequences to treat as circular (organelle) chromosomes}

\item{create_TxDb}{Create a \code{TxDb} object and a *Rannot object? It defaults to \code{TRUE}}

\item{annot_file}{specify an exact file name for the rds file created by this function, defaults to annotation_directory/basename(gtf)_Rannot}
}
\value{
a TxDb file and a *Rannot files are created in the specified \code{annotation_directory}.
In addition, a BSgenome object is forged, installed, and linked to the *Rannot object
}
\description{
This function processes a gtf file and a twobit file (created using faToTwoBit from ucsc tools: http://hgdownload.soe.ucsc.edu/admin/exe/ ) to create a com
prehensive set of genomic regions of interest in genomic and transcriptomic space (e.g. introns, UTRs, start/stop codons).
   In addition, by linking genome sequence and annotation, it extracts additional info, such as gene and transcript biotypes, genetic codes for different organelles, or chromosomes and transcripts lengths.
}
\details{
This function uses the \code{makeTxDbFromGFF} function to  create a TxDb object and extract
genomic regions and other info to a *Rannot R file; the \code{mapToTranscripts} and \code{mapFromTranscripts} functions are used to
map features to genomic or transcript-level coordinates. GTF file mist contain "exon" and "CDS" lines,
where each line contains "transcript_id" and "gene_id" values. Additional values such as "gene_biotype" or "gene_name" are also extracted.
Regarding sequences, the twobit file, together with input scientific and annotation names, is used to forge and install a
BSgenome package using the \code{forgeBSgenomeDataPkg} function.\cr\cr
The resulting GTF_annotation object (obtained after runnning \code{load_annotation}) contains:\cr\cr
\code{txs}: annotated transcript boundaries.\cr
\code{txs_gene}: GRangesList including transcript grouped by gene.\cr
\code{seqinfo}: indicating chromosomes and chromosome lengths.\cr
\code{start_stop_codons}: the set of annotated start and stop codon, with respective transcript and gene_ids.
reprentative_mostcommon,reprentative_boundaries and reprentative_5len represent the most common start/stop codon,
the most upstream/downstream start/stop codons and the start/stop codons residing on transcripts with the longest 5'UTRs\cr
\code{cds_txs}: GRangesList including CDS grouped by transcript.\cr
\code{introns_txs}: GRangesList including introns grouped by transcript.\cr
\code{cds_genes}: GRangesList including CDS grouped by gene.\cr
\code{exons_txs}: GRangesList including exons grouped by transcript.\cr
\code{exons_bins}: the list of exonic bins with associated transcripts and genes.\cr
\code{junctions}: the list of annotated splice junctions, with associated transcripts and genes.\cr
\code{genes}: annotated genes coordinates.\cr
\code{threeutrs}: collapsed set of 3'UTR regions, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{fiveutrs}: collapsed set of 5'UTR regions, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{ncIsof}: collapsed set of exonic regions of protein_coding genes, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{ncRNAs}: collapsed set of exonic regions of non_coding genes, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{introns}: collapsed set of intronic regions, with correspinding gene_ids. This set does not overlap exonic region.\cr
\code{intergenicRegions}: set of intergenic regions, defined as regions with no annotated genes on either strand.\cr
\code{trann}: DataFrame object including (when available) the mapping between gene_id, gene_name, gene_biotypes, transcript_id and transcript_biotypes.\cr
\code{cds_txs_coords}: transcript-level coordinates of ORF boundaries, for each annotated coding transcript. Additional columns are the same as as for the \code{start_stop_codons} object.\cr
\code{genetic_codes}: an object containing the list of genetic code ids used for each chromosome/organelle. see GENETIC_CODE_TABLE for more info.\cr
\code{genome}: the name of the forged BSgenome package, or an FaFile_Circ object. Loaded with \code{load_annotation} function.\cr
\code{stop_in_gtf}: stop codon, as defined in the annotation.\cr
}
\examples{
gtf_file <- system.file("extdata", "example.gtf",
package = "RiboseQC",mustWork = TRUE)
twobit_file <- system.file("extdata", "example.2bit",
                           package = "RiboseQC",mustWork = TRUE)
output_dir <- tempdir()
prepare_annotation_files(output_dir,twobit_file,gtf_file,
                   scientific_name = "Arabidopsis.thaliana",
                   annotation_name = "tair10.42" )

}
\seealso{
\code{\link{load_annotation}}, \code{\link{forgeBSgenomeDataPkg}}, \code{\link{makeTxDbFromGFF}}.
}
\author{
Lorenzo Calviello, \email{calviello.l.bio@gmail.com}
}
\keyword{RiboseQC}
