% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq_analysis.R
\name{RiboseQC_analysis}
\alias{RiboseQC_analysis}
\title{Perform a Ribo-seQC analysis}
\usage{
RiboseQC_analysis(
  annotation_file,
  bam_files,
  read_subset = TRUE,
  readlength_choice_method = "max_coverage",
  genome_seq = NULL,
  stranded = TRUE,
  normalize_cov = TRUE,
  chunk_size = 5000000L,
  write_tmp_files = TRUE,
  dest_names = NA,
  rescue_all_rls = FALSE,
  fast_mode = TRUE,
  create_report = TRUE,
  sample_names = NA,
  report_file = NA,
  extended_report = FALSE,
  pdf_plots = TRUE,
  offsets_df = NULL,
  BPPARAM = NULL,
  region_BPPARAM = NULL
)
}
\arguments{
\item{annotation_file}{Full path to the annotation file (*Rannot). Or, a vector with paths to one annotation file per bam file.}

\item{bam_files}{character vector containing the full path to the bam files}

\item{read_subset}{Select readlengths up to 99 percent of the reads, defaults to \code{TRUE}. Must be of length 1 or same length as bam_files.}

\item{readlength_choice_method}{Method used to subset relevant read lengths (see \code{choose_readlengths} function); defaults to 'max_coverage'. Must be of length 1 or same length as bam_files.}

\item{genome_seq}{An FaFile object, to be used instead of a BSgenome package}

\item{stranded}{are the analyzed libraries strand-specific? TRUE, FALSE or 'inverse'. Defaults to \code{TRUE}}

\item{normalize_cov}{export normalized (sum to 1 million) bedgraph files for coverage tracks? Defaults to \code{TRUE}}

\item{chunk_size}{the number of alignments to read at each iteration, defaults to 5000000, increase when more RAM is available. Must be between 10000 and 100000000}

\item{write_tmp_files}{Should output all the results (in *results_RiboseQC_all)? Defaults to \code{TRUE}. Must be of length 1 or same length as bam_files.}

\item{dest_names}{character vector containing the prefixes to use for the result output files. Defaults to same as \code{bam_files}}

\item{rescue_all_rls}{Set cutoff of 12 for read lengths ignored because of insufficient coverage. Defaults to \code{FALSE}. Must be of length 1 or same length as bam_files.}

\item{fast_mode}{Use only top 500 genes to build profiles? Defaults to \code{TRUE}. Must be of length 1 or same length as bam_files.}

\item{create_report}{Create an html report showing the RiboseQC analysis results. Defaults to \code{TRUE}}

\item{sample_names}{character vector containing the names for each sample analyzed (for the html report). Defaults to 'sample1', 'sample2' ...}

\item{report_file}{desired filename for for the html report file. Defaults to the first entry of \code{bam_files} followed by '.html'}

\item{extended_report}{creates a large html report including codon occupancy for each read length. Defaults to \code{FALSE}}

\item{pdf_plots}{creates a pdf file for each produced plot. Defaults to \code{TRUE}}

\item{offsets_df}{optionally input an offsets_df created externally, note that this does not yet fully integrate with stats such as 'proportion in phase' it will however alter the psites which are outputted. Must have cols 'read_length','cutoff'}

\item{BPPARAM}{optional \code{BiocParallelParam} object (e.g. \code{BiocParallel::MulticoreParam(4)}) used to
analyze several BAM files in parallel. Defaults to \code{NULL}: BAM files are analyzed one after the other.}

\item{region_BPPARAM}{optional \code{BiocParallelParam} object used to analyze each BAM file in parallel over
genomic windows (the BAM file must be indexed). Results are the same as with sequential processing.
Defaults to \code{NULL}: each BAM file is read sequentially in chunks of \code{chunk_size} alignments.
Use either \code{BPPARAM} or \code{region_BPPARAM}, not both, to avoid running more processes than cores.}
}
\value{
the function saves a 'results_RiboseQC_all' R file appended to the bam_files path including the complete list of outputs described here.
In addition, bedgraph files for coverage value and P_sites position is appended to the bam_files path, including also a summary of P_sites selection statistics,
a smaller 'results_RiboseQC' R file used for creating a dynamic html report, and a 'for_ORFquant' R object that can be used in the ORFquant pipeline.
}
\description{
This function loads annotation created by the prepare_annotation_files function, and analyzes a BAM file.
}
\details{
This function loads different genomic regions created in the \code{prepare_annotation_files} step,
separating features on different recognized organelles. The bam files is then analyzed in chunks to minimize RAM usage.\cr
The complete list of analysis and output is as follows:\cr\cr
\code{read_stats}: contains:\cr read length distribution (rld) per organelle, \code{positions} containes mapping statistics
on different genomic regions, \code{reads_pos1} contains 5' end mapping positions for each read, separated by read length.
\code{counts_cds_genes}: contains read mapping statistics on CDS regions of protein coding genes, including gene symbols, counts, RPKM and TPM values
\code{counts_all_genes}: is a similar object, but contains statistics on all annotated genes.
\code{reads_summary}: reports mapping statistics on different genomic regions and divided by read length and organelle.\cr\cr
\code{profiles_fivepr} contains:\cr
\code{five_prime_bins}: a DataFrame object (one for each read length and compartment) with signal values over 50 5'UTR bins,
100 CDS bins and 50 3'UTR bins; one representative transcript (reprentative_mostcommon) is selected for each gene. \code{five_prime_subcodon} containes a similar structure, but for 25nt downstream the Transcription
Start Site (TSS), 25nt upstream start codons, 33nt donwstream the start codon, 33nt in the middle of the ORF, 33nt upstream the stop codon,
25nt downstream the stop codon, and 25nt upstream the Transcription End Site (TES).\cr\cr
\code{selection_cutoffs} contains:\cr
\code{results_choice}: containing the calculated cutoffs and selected readlengths, together with \code{data} about the different
methods. \code{results_cutoffs} has statistics about calculated cutoffs, while \code{analysis_frame_cutoff} has extensive
statistics concerning cutoff calculations and read length selection, see \code{calc_cutoffs_from_profiles} for more details.\cr\cr
\code{P_sites_stats}: contains the list of calculated P_sites, from all reads (P_sites_all), uniquely mapping reads (P_sites_all_uniq),
or uniquely mapping reads with mismatches (P_sites_uniq_mm). \code{junctions} contains stastics on read mapping on annotated splice junctions.
coverage for entire reads (no 5'ends or P_sites-transformed) on different strands and for all and uniquely mapping reads are also calculated.\cr\cr
\code{profiles_P_sites} contains:\cr
\code{P_sites_bins}: profiles for each organelle and read length around binned transcript locations.\cr
\code{P_sites_subcodon}: profiles for each organelle and read length around transcript start/ends and ORF start/ends.\cr
\code{Codon_counts}: codon occurrences in the first 11 codons, middle 11 codons, and last 11 codons for each ORF.\cr
\code{P_sites_percodon}: P_sites counts on each codon, separated by ORF positions as described above. Values are separated by organelle and read length.\cr
\code{P_sites_percodon_ratio}: ratio of P_sites_percodon/Codon_counts, as a measure of P_site occupancy on each codon, divided again by organelle and read length, for different ORF positions.\cr\cr
\code{sequence_analysis}: contains a DataFrame object with the 50top mapping location in the genome, with the corresponding DNA sequence,
number of reads mapping (also in percentage of total n of reads), and genomic feature annotation.\cr\cr
\code{summary_P_sites}: contains a DataFrame object summarizing the P_sites calculation and read length selection, including statistics on percentage of total reads used.
}
\seealso{
\code{\link{prepare_annotation_files}}, \code{\link{calc_cutoffs_from_profiles}}, \code{\link{choose_readlengths}}, \code{\link{create_html_report}}.
}
\author{
Lorenzo Calviello, \email{calviello.l.bio@gmail.com}
}
\keyword{Ribo-seQC}
