% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{generate_rdata_list}
\alias{generate_rdata_list}
\title{Generate a list of R data objects}
\usage{
generate_rdata_list(input_files)
}
\arguments{
\item{input_files}{List of RData file paths generated by \code{RiboseQC_analysis}. \cr \cr
Example: \cr
\code{input_files <- c(sample1="//path//to//sample1", sample2="//path//to//sample2")}}
}
\value{
This function returns a list of loaded RData objects 
that were generated by \code{RiboseQC_analysis}.
}
\description{
This function generates a list of loaded RData files 
to be used during Ribo-seQC report generation.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
