% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_frame_dist_boxplot}
\alias{plot_frame_dist_boxplot}
\title{Plot frame coverage of 5'-site profiles}
\usage{
plot_frame_dist_boxplot(analysis_frame_cutoff, comp, sample = "",
  output_rds_path = "")
}
\arguments{
\item{analysis_frame_cutoff}{Object containing statistics on frame coverage
(per originating compartment and read length) generated by \code{RiboseQC_analysis}. \cr \cr
Example: \cr
\code{res_all$selection_cutoffs$analysis_frame_cutoff}}

\item{comp}{String for originating compartment \cr \cr
Check for available originating compartments in the data set using:
\code{names(res_all$profiles_fivepr$five_prime_subcodon)}}

\item{sample}{String; sample name (selected from the input names
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
This function returns a plot that can be integrated in the html report and 
that can be saved as RDS object file.
}
\description{
This function plots the frame coverage of 5'-site profiles, 
i.e. the fraction of reads (their 5' end) in frame 0, 1, and 2 (defined by start codon),
as boxplots (per-frame distributions over all transcripts)
for individual read lengths as well as for all read lengths summarized.
}
\details{
This plot is used in the Ribo-seQC report in section 4.2.1.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
