% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_frame_dist_boxplot_rmd}
\alias{plot_frame_dist_boxplot_rmd}
\title{Plot frame coverage of 5'-site profiles within the RMarkdown document}
\usage{
plot_frame_dist_boxplot_rmd(analysis_frame_cutoff, sample = "",
  output_rds_path = "")
}
\arguments{
\item{analysis_frame_cutoff}{Object containing statistics on frame coverage
(per originating compartment and read length) generated by \code{RiboseQC_analysis}. \cr \cr
Example: \cr
\code{res_all$selection_cutoffs$analysis_frame_cutoff}}

\item{sample}{String; sample name (selected from the input names
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
This function integrates plots in the html report.
}
\description{
This function generates iteratively all frame coverage boxplots
(iteration over originating compartments). \cr \cr
These plots are displayed in the Ribo-seQC report in section 4.2.1.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
