% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{plot_read_length_dist_by_biotype}
\alias{plot_read_length_dist_by_biotype}
\title{Plot read length and location distribution 
(distribution per biotype)}
\usage{
plot_read_length_dist_by_biotype(reads_summary, sample,
  output_rds_path = "")
}
\arguments{
\item{reads_summary}{\code{res_all$read_stats$reads_summary} generated by \code{RiboseQC_analysis} \cr \cr
List of DataFrames:
one DataFrame for each originating compartment,
each DataFrame contains read counts per biotype (rows) and read lengths (columns).}

\item{sample}{String; sample name (selected from the input names
given in the \code{input_sample_names} parameter of \code{create_html_report}).}

\item{output_rds_path}{String; full path to output folder for RDS object files created by this function.
Defaults to NOT save RDS; to save RDS, provide path to destination folder.}
}
\value{
This function returns a plot that can be integrated in the html report and 
that can be saved as RDS object file.
}
\description{
This funtion plots the read length distribution for each originating compartment as distribtion per biotype
for one input sample (displayed as read count and as read count fraction). \cr \cr
This plot is used in the Ribo-seQC report in section 3.1.
}
\seealso{
\code{\link{create_html_report}}
}
\author{
Dominique Sydow, \email{dominique.sydow@posteo.de}
}
\keyword{Ribo-seQC}
