% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseqc.R
\name{prepare_annotation_files}
\alias{prepare_annotation_files}
\title{Prepare comprehensive sets of annotated genomic features}
\usage{
prepare_annotation_files(annotation_directory, twobit_file = NULL, gtf_file,
  scientific_name = "Homo.sapiens", annotation_name = "genc25",
  export_bed_tables_TxDb = TRUE, forge_BSgenome = FALSE,
  genome_seq = NULL, circ_chroms = DEFAULT_CIRC_SEQS,
  create_TxDb = TRUE, annot_file = NULL)
}
\arguments{
\item{annotation_directory}{The target directory which will contain the output files}

\item{twobit_file}{Not used. Earlier versions forged a \code{BSgenome} package from this twobit file}

\item{gtf_file}{Full path to the annotation file in GTF format}

\item{scientific_name}{Not used. Earlier versions named the forged \code{BSgenome} package with it}

\item{annotation_name}{Not used. Earlier versions named the forged \code{BSgenome} package with it}

\item{export_bed_tables_TxDb}{Export coordinates and info about different genomic regions in the annotation_directory? It defaults to \code{TRUE}}

\item{forge_BSgenome}{Not used. Earlier versions forged and installed a \code{BSgenome} package from \code{twobit_file} when it was \code{TRUE};
this failed with current BSgenomeForge versions (ohlerlab/RiboseQC#21, #23). It defaults to \code{FALSE}}

\item{genome_seq}{FASTA file of the genome sequence (a path or an \code{FaFile}). Required; \code{twoBitToFa} of the UCSC tools
(http://hgdownload.soe.ucsc.edu/admin/exe/) makes one from a twobit file}

\item{circ_chroms}{Chromosomes to make circular in the genome sequence - defaults to DEFAULT_CIRC_SEQS}

\item{create_TxDb}{Create a \code{TxDb} object and a *Rannot object? It defaults to \code{TRUE}}

\item{annot_file}{specify an exact file name for the rds file created by this function, defaults to annotation_directory/basename(gtf)_Rannot}
}
\value{
a TxDb file and a *Rannot files are created in the specified \code{annotation_directory}.
}
\description{
This function processes a gtf file and a FASTA file of the genome to create a comprehensive set of genomic regions of interest in genomic and transcriptomic space (e.g. introns, UTRs, start/stop codons).
   In addition, by linking genome sequence and annotation, it extracts additional info, such as gene and transcript biotypes, genetic codes for different organelles, or chromosomes and transcripts lengths.
}
\details{
This function uses the \code{makeTxDbFromGFF} function to  create a TxDb object and extract
genomic regions and other info to a *Rannot R file; the \code{mapToTranscripts} and \code{mapFromTranscripts} functions are used to 
map features to genomic or transcript-level coordinates. GTF file mist contain "exon" and "CDS" lines,
where each line contains "transcript_id" and "gene_id" values. Additional values such as "gene_biotype" or "gene_name" are also extracted.
The genome sequence is read from the FASTA file \code{genome_seq}.\cr\cr
The resulting GTF_annotation object (obtained after runnning \code{load_annotation}) contains:\cr\cr
\code{txs}: annotated transcript boundaries.\cr
\code{txs_gene}: GRangesList including transcript grouped by gene.\cr
\code{seqinfo}: indicating chromosomes and chromosome lengths.\cr
\code{start_stop_codons}: the set of annotated start and stop codon, with respective transcript and gene_ids.
reprentative_mostcommon,reprentative_boundaries and reprentative_5len represent the most common start/stop codon,
the most upstream/downstream start/stop codons and the start/stop codons residing on transcripts with the longest 5'UTRs\cr
\code{cds_txs}: GRangesList including CDS grouped by transcript.\cr
\code{introns_txs}: GRangesList including introns grouped by transcript.\cr
\code{cds_genes}: GRangesList including CDS grouped by gene.\cr
\code{exons_txs}: GRangesList including exons grouped by transcript.\cr
\code{exons_bins}: the list of exonic bins with associated transcripts and genes.\cr
\code{junctions}: the list of annotated splice junctions, with associated transcripts and genes.\cr
\code{genes}: annotated genes coordinates.\cr
\code{threeutrs}: collapsed set of 3'UTR regions, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{fiveutrs}: collapsed set of 5'UTR regions, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{ncIsof}: collapsed set of exonic regions of protein_coding genes, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{ncRNAs}: collapsed set of exonic regions of non_coding genes, with correspinding gene_ids. This set does not overlap CDS region.\cr
\code{introns}: collapsed set of intronic regions, with correspinding gene_ids. This set does not overlap exonic region.\cr
\code{intergenicRegions}: set of intergenic regions, defined as regions with no annotated genes on either strand.\cr
\code{trann}: DataFrame object including (when available) the mapping between gene_id, gene_name, gene_biotypes, transcript_id and transcript_biotypes.\cr
\code{cds_txs_coords}: transcript-level coordinates of ORF boundaries, for each annotated coding transcript. Additional columns are the same as as for the \code{start_stop_codons} object.\cr
\code{genetic_codes}: an object containing the list of genetic code ids used for each chromosome/organelle. see GENETIC_CODE_TABLE for more info.\cr
\code{genome_package}: \code{NULL}. In annotations made by ORFquant with \code{forge_BSgenome=TRUE}, the name of the forged
BSgenome package, which \code{load_annotation} loads.\cr
\code{genome}: the genome sequence, an FaFile_Circ object. Loaded with \code{load_annotation} function.\cr
\code{stop_in_gtf}: stop codon, as defined in the annotation.\cr
}
\seealso{
\code{\link{load_annotation}}, \code{\link{makeTxDbFromGFF}}.
}
\author{
Lorenzo Calviello, \email{calviello.l.bio@gmail.com}
}
\keyword{Ribo-seQC}
